refactor(actions): collapse save-as-action into a single plain module
Replaces the kea logic + separate event/filter helpers with one module
at frontend/src/models/saveAsActionDialog.tsx. No state, no reducers,
no listeners — just plain functions: openSaveAsActionDialog,
saveActionFromFilter, saveActionFromEvent, plus the step/name/guard/
validator helpers.

Callers drop useMountedLogic + useActions. Deletes the old logic,
typegen, and the split scenes/activity/explore/saveActionFromEvent
module.

Generated-By: PostHog Code
Task-Id: 5cc6267b-3241-42ec-b029-997cd288ba9e
