+++ frontend/src/scenes/insights/filters/ActionFilter/ActionFilterRow/SaveAsActionBanner.tsx
@@ -1,11 +1,10 @@
-import { useActions, useMountedLogic } from 'kea'
-
 import { LemonBanner } from '@posthog/lemon-ui'
 import { PostHogCaptureOnViewed } from '@posthog/react'
 
 import { FlaggedFeature } from 'lib/components/FlaggedFeature'
 import { FEATURE_FLAGS } from 'lib/constants'
-import { saveAsActionLogic } from 'scenes/actions/saveAsActionLogic'
+
+import { saveActionFromFilter } from '~/models/saveAsActionDialog'
 
 import { LocalFilter } from '../entityFilterLogic'
 import { isAutocaptureFilterWithElements } from './saveAsActionUtils'
@@ -15,9 +14,6 @@ interface SaveAsActionBannerProps {
 }
 
 export function SaveAsActionBanner({ filter }: SaveAsActionBannerProps): JSX.Element | null {
-    useMountedLogic(saveAsActionLogic)
-    const { saveFromFilter } = useActions(saveAsActionLogic)
-
     if (!isAutocaptureFilterWithElements(filter)) {
         return null
     }
@@ -31,7 +27,7 @@ export function SaveAsActionBanner({ filter }: SaveAsActionBannerProps): JSX.Ele
                     dismissKey="autocapture-save-as-action-nudge"
                     action={{
                         children: 'Save as action',
-                        onClick: () => saveFromFilter(filter),
+                        onClick: () => saveActionFromFilter(filter),
                         'data-attr': 'autocapture-save-as-action',
                     }}
                 >
